rename classes and filenames to get rid of the old "instructions" naming
