+++ src/vs/workbench/contrib/chat/browser/media/chat.css
@@ -1005,52 +1005,52 @@ have to be updated for changes to the rules above, or to support more deeply nes
 /**
  * Styles for the `reusable prompts` attachment widget.
  */
-.chat-attached-context .chat-prompt-instructions-attachment {
+.chat-attached-context .chat-prompt-attachment {
 	display: flex;
 	gap: 4px;
 }
-.chat-attached-context .chat-prompt-instructions-attachment .codicon {
+.chat-attached-context .chat-prompt-attachment .codicon {
 	color: inherit;
 	text-decoration: none;
 }
-.chat-attached-context .chat-prompt-instructions-attachment .chat-implicit-hint {
+.chat-attached-context .chat-prompt-attachment .chat-implicit-hint {
 	opacity: 0.7;
 	font-size: .9em;
 	margin-top: -0.5px;
 }
-.chat-attached-context .chat-prompt-instructions-attachment.warning {
+.chat-attached-context .chat-prompt-attachment.warning {
 	color: var(--vscode-notificationsWarningIcon-foreground);
 }
-.chat-attached-context .chat-prompt-instructions-attachment.error {
+.chat-attached-context .chat-prompt-attachment.error {
 	color: var(--vscode-notificationsErrorIcon-foreground);
 }
-.chat-attached-context .chat-prompt-instructions-attachment.disabled {
+.chat-attached-context .chat-prompt-attachment.disabled {
 	border-style: dashed;
 	opacity: 0.75;
 }
-.chat-attached-context .chat-prompt-instructions-attachment .monaco-button {
+.chat-attached-context .chat-prompt-attachment .monaco-button {
 	border-left: 1px solid var(--vscode-chat-requestBorder, var(--vscode-input-background, transparent));
 	margin-left: 3px;
 }
-.chat-attached-context .chat-prompt-instructions-attachment.error .monaco-button,
-.chat-attached-context .chat-prompt-instructions-attachment.warning .monaco-button {
+.chat-attached-context .chat-prompt-attachment.error .monaco-button,
+.chat-attached-context .chat-prompt-attachment.warning .monaco-button {
 	border-left-color: currentColor;
 }
-.chat-attached-context .chat-prompt-instructions-attachment:focus .monaco-button {
+.chat-attached-context .chat-prompt-attachment:focus .monaco-button {
 	border-color: var(--vscode-focusBorder);
 }
-.chat-attached-context .chat-prompt-instructions-attachment .monaco-icon-label-container {
+.chat-attached-context .chat-prompt-attachment .monaco-icon-label-container {
 	margin-top: -0.1em;
 }
 /*
  * This overly-specific CSS selector is needed to beat priority of some
  * styles applied on the `.chat-attached-context-attachment` element.
  */
-.chat-attached-context .chat-prompt-instructions-attachment.error.implicit,
-.chat-attached-context .chat-prompt-instructions-attachment.warning.implicit {
+.chat-attached-context .chat-prompt-attachment.error.implicit,
+.chat-attached-context .chat-prompt-attachment.warning.implicit {
 	border: 1px solid currentColor;
 }
-.chat-attached-context .chat-prompt-instructions-attachment.implicit .monaco-button {
+.chat-attached-context .chat-prompt-attachment.implicit .monaco-button {
 	padding-left: 2px;
 	padding-right: 2px;
 }
@@ -1058,23 +1058,23 @@ have to be updated for changes to the rules above, or to support more deeply nes
  * If in one of the non-normal states, make sure the `main icon` of
  * the component has the same color as the component itself
  */
-.chat-attached-context .chat-prompt-instructions-attachment.error .monaco-icon-label::before,
-.chat-attached-context .chat-prompt-instructions-attachment.warning .monaco-icon-label::before,
-.chat-attached-context .chat-prompt-instructions-attachment.disabled .monaco-icon-label::before {
+.chat-attached-context .chat-prompt-attachment.error .monaco-icon-label::before,
+.chat-attached-context .chat-prompt-attachment.warning .monaco-icon-label::before,
+.chat-attached-context .chat-prompt-attachment.disabled .monaco-icon-label::before {
 	color: inherit;
 }
-.chat-attached-context .chat-prompt-instructions-attachment.disabled .monaco-icon-label::before {
+.chat-attached-context .chat-prompt-attachment.disabled .monaco-icon-label::before {
 	font-style: italic;
 }
-.chat-attached-context .chat-prompt-instructions-attachment.disabled:hover {
+.chat-attached-context .chat-prompt-attachment.disabled:hover {
 	opacity: 1;
 }
-.chat-attached-context .chat-prompt-instructions-attachment.disabled .chat-implicit-hint,
-.chat-attached-context .chat-prompt-instructions-attachment.disabled .label-name {
+.chat-attached-context .chat-prompt-attachment.disabled .chat-implicit-hint,
+.chat-attached-context .chat-prompt-attachment.disabled .label-name {
 	font-style: italic;
 	text-decoration: line-through;
 }
-.chat-attached-context .chat-prompt-instructions-attachment.disabled:focus {
+.chat-attached-context .chat-prompt-attachment.disabled:focus {
 	outline: none;
 	border-color: var(--vscode-focusBorder);
 }
